refactor(Utils): move three classes to the subsystem that owns them
TypeMapper moves to Type, with a deprecated subclass left behind; a custom type
resolver reaches it through AbstractTypeResolver. CollectingLogger moves to
Loggers beside DefaultLogger, and SpecificationWalker to Specification\Walker,
neither of which is documented surface.

Walker imports JsonPointer explicitly, which resolved by namespace before the
move.

pipeline.md states what Utils holds and what Contracts holds, next to the
existing note on directory layout.
