+++ src/Builder.php
@@ -9,8 +9,8 @@ namespace OpenApi;
 use OpenApi\Builder\Mode;
 use OpenApi\Builder\Result;
 use OpenApi\Contracts\CompilerInterface;
+use OpenApi\Loggers\CollectingLogger;
 use OpenApi\Utils\AttributeFactory;
-use OpenApi\Utils\CollectingLogger;
 use OpenApi\Utils\PipeInterface;
 use OpenApi\Utils\SourceScanner;
 use Psr\Log\LoggerInterface;
